Add low-side FET to properly disconnect battery on fault
What was done in v0.9 to disable charge/discharge by driving IO on
the charger circuit didn't really work out. Further, it had no way
of monitoring overcurrent during discharge.

Add a FED on the low side of the pack that can be controlled by the
protection circuit, since this is how it is supposed to work.

Fixes #6